Add new feature: Keep AVD alive

+++ utils/running_commands.py
@@ -1,6 +1,5 @@
 import subprocess
 import time 
-import keyboard
 import logging
 
 logger = logging.getLogger("cts_logger." + __name__)
@@ -16,20 +15,19 @@ class Commands:
             stdout=subprocess.PIPE,
             stderr=subprocess.PIPE,
             stdin=subprocess.DEVNULL,
-            text=True,  # Decode bytes to str
-            check=False
+            text=True  # Decode bytes to str
         )       
 
         stdout , stderr = process.communicate()
-        logger.info(f"Closing process")
-        logger.debug(f"Output : ", stdout)
-        logger.debug(f"Error : ", stderr)
+        logger.info(f"Closing process for command {cmd}")
+        logger.debug(f"Output : {stdout}")
+        logger.debug(f"Error : {stderr}")
         return stdout , stderr
 
     @staticmethod
-    def execute_timeout_cmd(cmd , timeout : int = 60):
+    def execute_timeout_cmd(cmd , timeout : int|None = 60):
         logger.info(f"Running command with timeout: {cmd}")
-        logger.debug(f"Timeout set {timeout}")
+        logger.info(f"Timeout set {timeout}")
         start_time = time.time()
         process = subprocess.Popen(
             cmd,
@@ -38,20 +36,21 @@ class Commands:
             stdout=subprocess.PIPE,
             stderr=subprocess.PIPE,
             stdin=subprocess.DEVNULL,
-            text=True,
-            check=False
+            text=True  # Decode bytes to str
         )
-        try:
-            while time.time() - start_time < timeout:
-                if keyboard.is_pressed('q'):
-                    logger.debug(f"User Cancellation")
-                    break
-                stdout, stderr = process.stdout.read(),process.stderr.read() # type: ignore
-                logger.debug(f"Output : ", stdout)
-                logger.debug(f"Error : ", stderr)
-        finally:
-            logger.info(f"Closing process")
-            process.terminate()
-            stdout, stderr = None , None
-        
-        return stdout, stderr
+        if timeout is None:
+            try:
+                logger.info("Not setting timeout for running, we wait 60 seconds for some outputs")
+                process.wait(60)
+            except subprocess.TimeoutExpired as e:
+                logger.info("Timeout for output reached, leaving process without closing it")
+        else:
+            try:
+                while time.time() - start_time < timeout:
+                    stdout, stderr = process.stdout.read(),process.stderr.read() # type: ignore
+                    logger.debug(f"Output : {stdout}")
+                    logger.debug(f"Error : {stderr}")
+            finally:
+                logger.info(f"Closing process for command {cmd}")
+                process.terminate()
+        return process
